moved tests to proper location
